+++ Mods ChoGGi/Show Tunnel Lines/items.lua
@@ -1,0 +1,14 @@
+return {
+	PlaceObj("ModItemOptionToggle", {
+		"name", "RemoveOnSelect",
+		"DisplayName", T(302535920011548, "Remove On Select"),
+		"Help", T(302535920011549, "Remove lines when selection changes."),
+		"DefaultValue", true,
+	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "RandomColours",
+		"DisplayName", T(302535920011550, "Random Colours"),
+		"Help", T(302535920011551, "Use a few colours, disable to only use white."),
+		"DefaultValue", true,
+	}),
+}